moved script handle functions
