+++ CMakeLists.txt
@@ -6,7 +6,8 @@ set(CMAKE_C_STANDARD 99)
 add_definitions(-DDEV)
 
 set(SOURCE_FILES src/main.c src/commandLine/CommandlineParser.h src/commandLine/CommandLineParser.c
-        src/utils/stringManip.h src/utils/stringManip.c src/utils/print.c src/utils/print.h src/Commands/Install.c src/Commands/Install.h)
+        src/utils/stringManip.h src/utils/stringManip.c src/utils/print.c src/utils/print.h src/Commands/Install.c
+        src/Commands/Install.h src/build_parser/coldbrew/scriptHandler.c src/build_parser/coldbrew/scriptHandler.h)
 
 #if we need hashmap string->int stuff uncomment the line below
 #set(SOURCE_FILES ${SOURCE_FILES} src/hashMap/hash_map_string_int.h src/hashMap/hash_map_string_int.c src/hashMap/hash.h src/hashMap/hash.c)
