+++ src/Commands/Install.c
@@ -9,6 +9,8 @@
 #include <stdlib.h>
 #include "Install.h"
 #include "../utils/print.h"
+#include "../build_parser/coldbrew/scriptHandler.h"
+
 
 const short SHABANG = 0x2321;
 
@@ -114,33 +116,6 @@ int get_first_int(FILE *file)
 	fread(ret, 1, 1, file);
 	//return
 	return *ret;
-}
-
-//this should be moved to a build script file
-void build_blob(struct target* targ){
-	//create tmp dir if needed
-	if (strstr(targ->tmp_dir, "/tmp/coldbrew.") == NULL){
-		char tmp_dir_template[256];
-		//create a template for the temp dir, this will be where we build the package and such.
-		strcat(tmp_dir_template, "/tmp/coldbrew.");
-		strcat(tmp_dir_template, targ->name);
-		strcat(tmp_dir_template, ".XXXXXX");
-		targ->tmp_dir = mkdtemp(tmp_dir_template);
-		dbprintf(DEBUG, "target: %s, tempdir: %s\n", targ->name, targ->tmp_dir);
-	}
-	//parse script
-	parse_script(targ->blob_script_loc);
-	//extract srcs
-	//run prep()
-	//run build()
-	//run check()
-	//run package()
-	//compress package
-}
-
-//this should be moved to a build script file with the above funct
-void parse_script(char* script){
-
 }
 
 void install_blob(struct target* targ) {
