+++ src/Commands/Install.h
@@ -19,10 +19,8 @@ char * download_file(struct target targ, char *tmp_dir);
 
 int get_first_int(FILE* file);
 
-void build_blob(struct target* targ);
 
 void install_blob(struct target* targ);
 
-void parse_script(char* script);
 
 #endif //COLDBREW_INSTALL_H
