+++ src/build_parser/coldbrew/scriptHandler.h
@@ -1,0 +1,14 @@
+//
+// Created by Garrett Battaglia on 12/27/17.
+//
+
+#ifndef COLDBREW_SCRIPTHANDLER_H
+#define COLDBREW_SCRIPTHANDLER_H
+
+#include "../../commandLine/CommandlineParser.h"
+
+void parse_script(char* script);
+
+void build_blob(struct target* targ);
+
+#endif //COLDBREW_SCRIPTHANDLER_H
